refactor api graph loader
